[base] Make the narrowing in `misalignment32_of` explicit
Fix `-Wshorten-cap-to-int` warning for CHERIoT. `misalignment32_of`
really intends to return an offset, not a pointer, so the explicit cast
reflects that and silences the warning.

+++ sw/device/lib/base/memory.h
@@ -55,7 +55,7 @@ extern "C" {
  */
 OT_WARN_UNUSED_RESULT
 inline ptrdiff_t misalignment32_of(uintptr_t addr) {
-  return addr % alignof(uint32_t);
+  return (ptrdiff_t)(addr % alignof(uint32_t));
 }
 
 /**
